test(typescript): drop tsd for vitest type testing API
Now that we are using vitest for tests anyway, it comes with an
experimental type testing API which is worth using.

This doesn't affect users, but as maintainers, this drops the total
dependencies (including transitive dependencies) from 610 to 486. So
ideally that means 124 fewer things that can go wrong. 😅

I did also explore if we truly needed type testing, or if tsc would be
enough, but indeed, a real type testing library would still be useful as
tsc has no way to evaluate exact types without plumbing. As in, tsc
natively only checks if types overlap or satisfy each other, but not if
the types precisely match the assertion.

Useful article:
https://www.totaltypescript.com/how-to-test-your-types
